New module added, main.js have been corrected
